+++ packages/injected-script/sdk/index.ts
@@ -1,14 +1,14 @@
-import { CustomEventId, decodeEvent } from '../shared/index.js'
+import { BrowserProvider } from './Browser.js'
 import { Coin98Provider, Coin98ProviderType } from './Coin98.js'
 import { CoinbaseProvider } from './Coinbase.js'
 import { OKXProvider } from './OKX.js'
 import { PhantomProvider } from './Phantom.js'
 import { SolflareProvider } from './Solflare.js'
 import { MetaMaskProvider } from './MetaMask.js'
-import { sendEvent, rejectPromise, resolvePromise } from './utils.js'
 import { OperaProvider } from './Opera.js'
 import { CloverProvider } from './Clover.js'
-import { BrowserProvider } from './Injected.js'
+import { sendEvent, rejectPromise, resolvePromise } from './utils.js'
+import { CustomEventId, decodeEvent } from '../shared/index.js'
 
 export type { EthereumProvider, InternalEvents } from '../shared/index.js'
 export { InjectedProvider } from './Base.js'
