+++ packages/web3-providers/src/Web3/EVM/providers/Browser.ts
@@ -3,18 +3,18 @@ import { type ChainId, ProviderType, type Web3, type Web3Provider } from '@maskn
 import { BaseInjectedProvider } from './BaseInjected.js'
 import type { WalletAPI } from '../../../entry-types.js'
 
-export class InjectedProvider
+export class BrowserProvider
     extends BaseInjectedProvider
     implements WalletAPI.Provider<ChainId, ProviderType, Web3Provider, Web3>
 {
     constructor() {
-        super(ProviderType.Injected, injectedBrowserProvider)
+        super(ProviderType.Browser, injectedBrowserProvider)
     }
 
     protected override onAccountsChanged(accounts: string[]): void {
         if (accounts.length) this.emitter.emit('accounts', accounts)
         // disconnection will trigger an empty accounts list
-        else this.emitter.emit('disconnect', ProviderType.Injected)
+        else this.emitter.emit('disconnect', ProviderType.Browser)
     }
 
     override async disconnect(): Promise<void> {
